fix algorithm

+++ Plugin/include/dynmesh.hpp
@@ -7,10 +7,10 @@ typedef float SDFValue;
 
 typedef Eigen::Vector3i Position;
 typedef Eigen::Vector2i Edge;
-typedef size_t Id;
+typedef int Id;
 
 typedef Eigen::Vector3f VertexPosition;
-typedef size_t VertexId;
+typedef int VertexId;
 
 extern "C" void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API naive_surface_nets(
-    VertexPosition *vertices, VertexId *triangles, SDFValue *sdf, size_t size);
+    VertexPosition *vertices, VertexId *triangles, SDFValue *sdf, int size);
